Попытка убрать конфликты для добавления ветки по удалению фильмов и пользователей

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -73,6 +73,11 @@ public class UserService {
     }
 
     public long deleteUser(long id) {
-        return userStorage.delete(id);
+        if (id > 0) {
+            return userStorage.delete(id);
+        } else {
+            throw new NotFoundException("Пользователь с id " + id + " не существует");
+        }
+
     }
 }
